requirements.txt
requests==2.32.4
youtube-transcript-api==1.1.1
openai==1.12.0
pydantic==2.5.0
python-dotenv==1.1.1